WEBAPP-400: Use PayloadDataType
