+++ src/modules/app/reducers/index.js
@@ -15,9 +15,11 @@ import {
 import { handleActions } from 'redux-actions'
 import { startFetchActionName } from '../../app/actions/startFetchAction'
 import { finishFetchActionName } from '../../app/actions/finishFetchAction'
-import type { ActionType } from '../../app/createReduxStore'
 import type { Reducer } from 'redux'
 import type { StateType } from '../StateType'
+import type { StartFetchActionType } from '../../app/actions/startFetchAction'
+import type { FinishFetchActionType } from '../../app/actions/finishFetchAction'
+import type { PayloadDataType } from '../PayloadDataType'
 
 /**
  * Contains all endpoints which are defined in {@link './endpoints/'}
@@ -34,10 +36,11 @@ const endpoints = [
   poisEndpoint
 ]
 
-export const startFetchReducer = <T> (oldPayload?: Payload<T>, action: ActionType<T>): Payload<T> =>
-  action.payload
+export const startFetchReducer = <T: PayloadDataType> (oldPayload?: Payload<T>, action: StartFetchActionType<T>
+): Payload<T> => action.payload
 
-export const finishFetchReducer = <T> (oldPayload?: Payload<T>, action: ActionType<T>): Payload<T> => {
+export const finishFetchReducer = <T: PayloadDataType> (oldPayload?: Payload<T>, action: FinishFetchActionType<T>
+): Payload<T> => {
   if (!oldPayload) {
     return action.payload
   }
@@ -54,7 +57,8 @@ export const finishFetchReducer = <T> (oldPayload?: Payload<T>, action: ActionTy
 
 const defaultState = new Payload(false)
 
-const reducers: { [actionName: string]: Reducer<StateType, *> } = endpoints.reduce(
+type ReducerType = Reducer<StateType, StartFetchActionType<PayloadDataType> | FinishFetchActionType<PayloadDataType>>
+const reducers: { [actionName: string]: ReducerType } = endpoints.reduce(
   (result, endpoint) => {
     const name = endpoint.stateName
     result[name] = handleActions(
